third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-sans-fieldset-display-ruby-expected.txt

This is a testharness.js-based test.
[FAIL] .ruby
  assert_equals: .ruby display expected "ruby" but got "block ruby"
[FAIL] .rt
  assert_equals: .rt display expected "ruby-text" but got "inline-block"
Harness: the test ran to completion.
